Add database deletion and migration in development mode
The code now includes changes to enable database deletion and migration when the system is running in development mode. The implemented changes make use of the LeadManagementSystemEFCoreDbContext and Microsoft.EntityFrameworkCore for managing the database setup. This will assist programmers in testing and developing new features by allowing them to work with the up-to-date database schema.